feat: add JWT support
GraphQL now uses JWT tokens for authentication. Do the following:

```
edgeca config graphql --port 9000
edgeca server
```

Note the JWT token output when the server starts. Then go to

[GraphQL playground](http://localhost:9000) and try the following query:

```
mutation {
  createCertificate(input: {commonName: "localhost",organization:"ACME inc"}){
    certificate,
    key,
    expiry
  }
}
```

This will return in an error as you are not authenticated. Authenticate by clicking on "HTTP HEADERS" in the bottom pane,
and enter the JWT token. For example:

```
{"Authorization": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE2MzM0NTM5OTIsImlzcyI6IkVkZ2VDQSIsInN1YiI6InVzZXIifQ.js-Iq8gJFD9BwNGWAwnor4sz7KHYV5z2Yb4FdfwT9FLRpNpDALqELruB6V1wiWYqi2PbfwdT0ksitZuz9NTGQcnlYPdEt8TSizOdzl7ABoweagkF7OEwEOgTN1S2waD9vs4AwOGtqVPC8gXVgSV3UtG68F6GA8FsJtKdXeJzUIU"}
```

Run the query again and you should now get the expected result.